Fix gap property type in SpriteAtlasJson and update sprite rendering logic to account for frame gaps

+++ Galaga/Interfaces/SpriteAtlasJson.cs
@@ -9,7 +9,7 @@ namespace Galaga.Interfaces
         public string TexturePath { get; set; }
 
         [JsonPropertyName("gap")]
-        public long Gap { get; set; }
+        public float Gap { get; set; }
 
         [JsonPropertyName("sprites")]
         public Sprites[] Sprites { get; set; }
